Update README.md
This project is a Movie Recommendation System built using Python and data science techniques. The goal is to suggest movies that users are likely to enjoy, based on similarity in content and user preferences.

Instead of just picking random recommendations, the system analyzes features such as genres, keywords, cast, and crew to build connections between movies. It then ranks and recommends films that share the most in common.